+++ source/envfileparser/__init__.py
@@ -1,1 +1,4 @@
-from .env_file_parser import get_env, get_envs
+from .get_env import get_env
+from .get_envs import get_envs
+
+__all__ = ['get_env', 'get_envs']
